make it more similar to csharp version
