+++ python/README.md
@@ -1,0 +1,14 @@
+Fantasy Battle in Python
+========================
+
+For exercise instructions see (top level README)[../README.md]
+
+Suggestion: create a python virtual environment for this project. See the [documentation](https://docs.python.org/3/library/venv.html)
+
+install requirements:
+
+    pip install --upgrade -r requirements.txt
+
+run tests:
+
+    python -m pytest
