--- python/src/Damage.py
@@ -1,5 +1,0 @@
-class Damage:
-    amount: int
-
-    def __init__(self, amount: int) -> None:
-        self.amount = amount
