--- python/src/Inventory.py
@@ -1,3 +1,0 @@
-class Inventory:
-    def __init__(self, equipment):
-        self.equipment = equipment
