--- python/src/Stats.py
@@ -1,3 +1,0 @@
-class Stats:
-    def __init__(self, strength: int) -> None:
-        self.strength = strength
